RavenDB-17373 updated MemoryCache with latest sources
